requirements.txt
openai>=1.0.0
openai-whisper>=20231117
pydub>=0.25.1
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.2.0
transformers>=4.30.0
torch>=2.0.0
python-dotenv>=1.0.0
tqdm>=4.65.0 